travel_agent/requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
streamlit>=1.35.0
streamlit-folium>=0.20.0
tavily-python>=0.3.0
folium>=0.16.0
fpdf2>=2.7.0
plotly>=5.22.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
geopy>=2.4.0
